fix(contract): audit H1/M1/L1/L4/L5 — deadline gates, constructor-pinned admin, persistentCommit sealed submissions
H1 (stake-trap, mechanically confirmed): createWager requires a future
deadline; acceptWager and submitWorkout reject once the deadline passes
(blockTimeLt) — previously a past-deadline wager let a challenger
forfeit-take an accepting opponent's escrowed stake with no chance to
submit. M1: admin identity pinned by the constructor at deploy (the old
first-call-wins registerAdmin left a front-run race on the notary
registry). L1: sealed submissions are persistentCommit<Field>(value, rand)
with Bytes<32> rands — Wager submissions become Maybe<Bytes<32>>; private
state + witnesses carry bytes. L4: self-challenges rejected. L5: stake
capped at 2^63-1 so 2*stake always fits Uint<64> at settle.

StrideSim gains block-time control (sim.now) for deterministic
time-travel; the AUDIT exploit test flips to the rejection; new tests for
accept/submit-after-deadline, settle-after-grace, self-challenge, stake
cap and stranger-seizing-admin. Suite 49/49.

+++ packages/contract/src/witnesses.ts
@@ -44,14 +44,14 @@ export const witnesses = {
 
   getSubmissionRand: ({
     privateState,
-  }: WitnessBase): [PrivateState, bigint] => [
+  }: WitnessBase): [PrivateState, Uint8Array] => [
     privateState,
     privateState.submissionRand,
   ],
 
   getWagerOpenings: ({
     privateState,
-  }: WitnessBase): [PrivateState, [bigint, bigint, bigint, bigint]] => [
+  }: WitnessBase): [PrivateState, [bigint, Uint8Array, bigint, Uint8Array]] => [
     privateState,
     privateState.wagerOpenings,
   ],
